+++ packages/rfw/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.19
+* Add `DropdownButton` and `ClipRRect` widgets to rfw widget library.
+
+## 1.0.18
+* Exposes `WidgetLibrary`s registered in `Runtime`.
+* Exposes widgets map in `LocalWidgetLibrary`.
+
 ## 1.0.17
 
 * Adds support for tracking source locations of `BlobNode`s and